Mostly finished functionality to add trials

+++ MogulMoves/app/src/main/java/com/example/mogulmoves/TrialSerializer.java
@@ -77,7 +77,7 @@ public class TrialSerializer implements Serializer<Trial> {
 
         }  else {
 
-            float measurement = (int) (long) map.get("measurement");
+            float measurement = (float) (double) map.get("measurement");
             trial = new MeasureTrial(id, owner, measurement);
         }
